Filled field float

+++ server/actions/patient.go
@@ -14,7 +14,7 @@ type BloodTestFilledField struct {
 	BloodTestFieldId uint                 `json:"blood_test_field_id"`
 	Name             string               `json:"name"`
 	Unit             models.BlootTestUnit `json:"unit"`
-	ValueNumber      int                  `json:"value_number"`
+	ValueNumber      float64              `json:"value_number"`
 	ValueString      string               `json:"value_string"`
 }
